requirements.txt
# Polymarket Streak-Reversal Martingale Bot
# pip install -r requirements.txt --break-system-packages

# Polymarket CLOB SDK
py-clob-client>=0.10.0

# Web3 + Polygon
web3>=6.0.0

# WebSocket (data feed)
websocket-client>=1.6.0

# HTTP
requests>=2.28.0

# Telegram
python-telegram-bot>=20.0

# Env
python-dotenv>=1.0.0

# Utility
httpx>=0.24.0

# Dashboard
rich>=13.0.0
